bug 1063565 - Remove doc-style tests
The documentation-style integration tests in test_viewset_* replicate
the tests in docs/doc_cases.json, and can be removed. Some
integration-style tests remain, for code coverage.

test_view_serializers.py (renamed from test_viewset_viewfeatureset.py)
is mostly unchanged. view_serializers.py will undergo a rewrite w/ bug
1153288, and the tests will also change.
