+++ webplatformcompat/tasks.py
@@ -3,9 +3,6 @@ from django.conf import settings
 
 from .cache import Cache
 
-DRF_INSTANCE_CACHE_POPULATE_COLD = getattr(
-    settings, 'DRF_INSTANCE_CACHE_POPULATE_COLD', True)
-
 
 @shared_task(ignore_result=True)
 def update_cache_for_instance(
@@ -14,6 +11,8 @@ def update_cache_for_instance(
     cache = Cache()
     invalid = cache.update_instance(
         model_name, instance_pk, instance, version, update_only=update_only)
+    DRF_INSTANCE_CACHE_POPULATE_COLD = getattr(
+        settings, 'DRF_INSTANCE_CACHE_POPULATE_COLD', True)
     for invalid_name, invalid_pk, invalid_version in invalid:
         update_cache_for_instance.delay(
             invalid_name, invalid_pk, version=invalid_version,
