+++ webplatformcompat/tests/test_models.py
@@ -4,10 +4,38 @@
 import mock
 import unittest
 
+from webplatformcompat.history import Changeset
 from webplatformcompat.models import (
     Browser, Feature, Maturity, Section, Specification, Support, Version,
     post_save_update_cache)
 from .base import TestCase
+
+
+class TestManager(TestCase):
+    def setUp(self):
+        self.login_user()
+        changeset = Changeset.objects.create(user=self.user)
+        self.patcher1 = mock.patch(
+            'webplatformcompat.history.HistoricalRecords'
+            '.get_history_changeset')
+        self.mocked_get_history_changeset = self.patcher1.start()
+        self.mocked_get_history_changeset.return_value = changeset
+        self.patcher2 = mock.patch(
+            'webplatformcompat.tasks.update_cache_for_instance')
+        self.mocked_update_cache = self.patcher2.start()
+
+    def tearDown(self):
+        self.patcher2.stop()
+        self.patcher1.stop()
+
+    def test_create(self):
+        browser = Browser.objects.create()
+        self.mocked_update_cache.assertCalledOnce(
+            'Browser', browser.id, browser)
+
+    def test_create_delay(self):
+        Browser.objects.create(_delay_cache=True)
+        self.mocked_update_cache.assertNotCalled()
 
 
 class TestBrowser(unittest.TestCase):
@@ -87,6 +115,11 @@ class TestSaveSignal(unittest.TestCase):
         post_save_update_cache(Browser, self.browser, created=True, raw=False)
         self.mocked_update_cache.assertCalledOnce('Browser', 666, self.browser)
 
+    def test_create_delayed(self):
+        self.browser._delay_cache = True
+        post_save_update_cache(Browser, self.browser, created=True, raw=False)
+        self.mocked_update_cache.assertNotCalled()
+
 
 class TestM2MChangedSignal(TestCase):
     def setUp(self):
@@ -94,7 +127,7 @@ class TestM2MChangedSignal(TestCase):
             'webplatformcompat.tasks.update_cache_for_instance')
         self.login_user()
         self.mocked_update_cache = self.patcher.start()
-        self.maturity = self.create(Maturity, slug='Foo')
+        self.maturity = self.create(Maturity, slug='foo')
         self.specification = self.create(Specification, maturity=self.maturity)
         self.section = self.create(Section, specification=self.specification)
         self.feature = self.create(Feature)
@@ -111,6 +144,11 @@ class TestM2MChangedSignal(TestCase):
         self.mocked_update_cache.assertCalledOnce(
             'Feature', self.feature.pk, self.feature)
 
+    def test_add_section_to_feature_delayed(self):
+        self.feature._delay_cache = True
+        self.feature.sections.add(self.section)
+        self.mocked_update_cache.assertNotCalled()
+
     def test_add_feature_to_section(self):
         self.section.features.add(self.feature)
         self.mocked_update_cache.assertCalledOnce(
@@ -120,3 +158,25 @@ class TestM2MChangedSignal(TestCase):
         self.section.features.clear()
         self.mocked_update_cache.assertCalledOnce(
             'Feature', self.feature.pk, self.feature)
+
+
+class TestDeleteSignal(TestCase):
+    def setUp(self):
+        self.patcher = mock.patch(
+            'webplatformcompat.tasks.update_cache_for_instance')
+        self.login_user()
+        self.mocked_update_cache = self.patcher.start()
+        self.maturity = self.create(Maturity, slug='foo')
+
+    def tearDown(self):
+        self.patcher.stop()
+
+    def test_delete(self):
+        self.maturity.delete()
+        self.mocked_update_cache.assertCalledOnce(
+            'Maturity', self.maturity.pk, self.maturity)
+
+    def test_delete_delayed(self):
+        self.maturity._delay_cache = True
+        self.maturity.delete()
+        self.mocked_update_cache.assertNotCalled()
