--- webplatformcompat/tests/test_viewset_historicalfeature.py
@@ -1,171 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""Tests for `web-platform-compat.viewsets.HistoricalFeatureViewSet` class."""
-from __future__ import unicode_literals
-from datetime import datetime
-from json import loads
-from pytz import UTC
-
-from django.core.urlresolvers import reverse
-
-from webplatformcompat.models import Feature, Maturity, Section, Specification
-
-from .base import APITestCase
-
-
-class TestHistoricalFeatureViewset(APITestCase):
-    def setUp(self):
-        self.user = self.login_user()
-
-    def test_get(self):
-        feature = self.create(
-            Feature, slug="the_feature", name={"en": "The Feature"},
-            _history_user=self.user,
-            _history_date=datetime(2014, 10, 1, 14, 25, 14, 955097, UTC))
-        history = feature.history.all()[0]
-        url = reverse('historicalfeature-detail', kwargs={'pk': history.pk})
-        response = self.client.get(url, HTTP_ACCEPT="application/vnd.api+json")
-        self.assertEqual(200, response.status_code, response.data)
-
-        expected_data = {
-            'id': history.pk,
-            'date': self.dt_repr(feature._history_date),
-            'event': 'created',
-            'changeset': history.history_changeset_id,
-            'feature': feature.pk,
-            'features': {
-                'id': str(feature.id),
-                'slug': 'the_feature',
-                'mdn_uri': None,
-                'experimental': False,
-                'standardized': True,
-                'stable': True,
-                'obsolete': False,
-                'name': {'en': 'The Feature'},
-                'links': {
-                    'parent': None,
-                    'sections': [],
-                    'history_current': str(history.pk),
-                }
-            },
-        }
-        self.assertDataEqual(expected_data, response.data)
-        expected_json = {
-            'historical_features': {
-                'id': str(history.pk),
-                'date': self.dt_json(feature._history_date),
-                'event': 'created',
-                'features': {
-                    'id': str(feature.id),
-                    'slug': 'the_feature',
-                    'mdn_uri': None,
-                    'experimental': False,
-                    'standardized': True,
-                    'stable': True,
-                    'obsolete': False,
-                    'name': {'en': 'The Feature'},
-                    'links': {
-                        'parent': None,
-                        'sections': [],
-                        'history_current': str(history.pk),
-                    }
-                },
-                'links': {
-                    'feature': str(feature.pk),
-                    'changeset': str(history.history_changeset_id),
-                },
-            },
-            'links': {
-                'historical_features.feature': {
-                    'href': (
-                        'http://testserver/api/v1/features/'
-                        '{historical_features.feature}'),
-                    'type': 'features'
-                },
-                'historical_features.changeset': {
-                    'href': (
-                        'http://testserver/api/v1/changesets/'
-                        '{historical_features.changeset}'),
-                    'type': 'changesets'
-                },
-            }
-        }
-        actual_json = loads(response.content.decode('utf-8'))
-        self.assertDataEqual(expected_json, actual_json)
-
-    def test_get_with_sections(self):
-        feature = self.create(
-            Feature, slug="the_feature", name={"en": "The Feature"},
-            _history_user=self.user,
-            _history_date=datetime(2014, 10, 1, 14, 25, 14, 955097, UTC))
-        maturity = self.create(Maturity, slug='Bar')
-        specification = self.create(Specification, maturity=maturity)
-        section = self.create(Section, specification=specification)
-        feature.sections.add(section)
-        feature.save()
-        self.assertEqual(2, len(feature.history.all()))
-        history = feature.history.all()[0]
-        url = reverse('historicalfeature-detail', kwargs={'pk': history.pk})
-        response = self.client.get(url, HTTP_ACCEPT="application/vnd.api+json")
-        self.assertEqual(200, response.status_code, response.data)
-
-        expected_data = {
-            'id': history.pk,
-            'date': self.dt_repr(feature._history_date),
-            'event': 'changed',
-            'changeset': history.history_changeset_id,
-            'feature': feature.pk,
-            'features': {
-                'id': str(feature.id),
-                'slug': 'the_feature',
-                'mdn_uri': None,
-                'experimental': False,
-                'standardized': True,
-                'stable': True,
-                'obsolete': False,
-                'name': {'en': 'The Feature'},
-                'links': {
-                    'parent': None,
-                    'sections': [str(section.pk)],
-                    'history_current': str(history.pk),
-                }
-            },
-        }
-        self.assertDataEqual(expected_data, response.data)
-
-    def test_filter_by_id(self):
-        parent = self.create(Feature, slug="parent-feature")
-        feature = self.create(
-            Feature, slug="a-feature", parent=parent,
-            _history_user=self.user,
-            _history_date=datetime(2014, 10, 1, 14, 29, 33, 22803, UTC))
-        history = feature.history.all()[0]
-        url = reverse('historicalfeature-list')
-        response = self.client.get(url, {'id': feature.id})
-        expected_data = {
-            'count': 1,
-            'previous': None,
-            'next': None,
-            'results': [{
-                'id': history.pk,
-                'date': self.dt_repr(feature._history_date),
-                'event': 'created',
-                'changeset': history.history_changeset_id,
-                'feature': feature.pk,
-                'features': {
-                    'id': str(feature.pk),
-                    'slug': 'a-feature',
-                    'mdn_uri': None,
-                    'experimental': False,
-                    'standardized': True,
-                    'stable': True,
-                    'obsolete': False,
-                    'name': None,
-                    'links': {
-                        'parent': str(parent.id),
-                        'sections': [],
-                        'history_current': str(history.pk),
-                    }
-                },
-            }]}
-        self.assertDataEqual(expected_data, response.data)
